Partial solution: Change calculator
Most of the way through this one. Asks for a total purchase cost, and
the amount used to pay for the purchase. Returns the total change given,
and a breakdown of the bills and coins it might be provided as.

Currently this works for the bills portion. Coins is not. Going to need
to figure out how to account for division when `1 > value > 0`, I think.